+++ mobile/android/android-components/components/browser/engine-gecko/src/main/java/mozilla/components/browser/engine/gecko/ipprotection/GeckoIPProtectionHandler.kt
@@ -120,4 +120,8 @@ internal class GeckoIPProtectionHandler(
     override fun notifyAccountStatus(signedIn: Boolean) {
         runtime.ipProtectionController.notifySignInStateChanged(signedIn)
     }
+
+    override fun setLocation(code: String?) {
+        runtime.ipProtectionController.setLocation(code)
+    }
 }