+++ mobile/shared/components/geckoview/GeckoViewStartup.sys.mjs
@@ -217,6 +217,7 @@ export class GeckoViewStartup {
               "GeckoView:IPProtection:Activate",
               "GeckoView:IPProtection:Deactivate",
               "GeckoView:IPProtection:Enroll",
+              "GeckoView:IPProtection:SetLocation",
             ],
           });